Update caveman config generator to use generator utils
