+++ utils/draw_topo.py
@@ -2,6 +2,9 @@
 Program for drawing network from json file
 input: relative path to json file
 Graphviz library must be installed
+
+NOTE: this file currently only works for sequential simulation files.
+If your JSON file contains parallel simulation information, please remove before use.
 """
 
 import argparse
