+++ utils/json_config_generators/config_generator_caveman.py
@@ -28,6 +28,8 @@ import argparse
 import json
 import pandas as pd
 
+from generator_utils import *
+
 from sequence.topology.topology import Topology
 from sequence.topology.router_net_topo import RouterNetTopo
 
@@ -70,91 +72,25 @@ nx.relabel_nodes(graph, mapping, copy=False)
 
 output_dict = {}
 
+# get node names, processes
 node_procs = {}
-router_names = []
-
 if args.nodes:
-    # TODO: add length/proc assertions
-    df = pd.read_csv(args.nodes)
-    for name, group in zip(df['name'], df['group']):
-        node_procs[name] = group
-        router_names.append(name)
+    node_procs = get_node_csv(args.nodes)
 else:
-    PROC_NUM = 1
-    if args.parallel:
-        PROC_NUM = int(args.parallel[2])
-    group_size = NODE_NUM / PROC_NUM
-    for i in range(NODE_NUM):
-        name = router_name_func(i)
-        node_procs[name] = int(i // group_size)
-        router_names.append(name)
-
+    node_procs = generate_node_procs(args.parallel, NODE_NUM, router_name_func)
 router_names = list(node_procs.keys())
-nodes = [{Topology.NAME: name,
-          Topology.TYPE: RouterNetTopo.QUANTUM_ROUTER,
-          Topology.SEED: i,
-          RouterNetTopo.MEMO_ARRAY_SIZE: args.memo_size,
-          RouterNetTopo.GROUP: node_procs[name]}
-         for i, name in enumerate(router_names)]
+nodes = generate_nodes(node_procs, router_names, args.memo_size)
 
-cchannels = []
-qchannels = []
-bsm_nodes = []
-for i, node_pair in enumerate(graph.edges):
-    node1, node2 = node_pair
-    bsm_name = bsm_name_func(node1, node2)
-    bsm_node = {Topology.NAME: bsm_name,
-                Topology.TYPE: RouterNetTopo.BSM_NODE,
-                Topology.SEED: i,
-                RouterNetTopo.GROUP: node_procs[node1]}
-    bsm_nodes.append(bsm_node)
-
-    for node in node_pair:
-        qchannels.append({Topology.SRC: node,
-                          Topology.DST: bsm_name,
-                          Topology.DISTANCE: args.qc_length * 500,
-                          Topology.ATTENUATION: args.qc_atten})
-
-    for node in node_pair:
-        cchannels.append({Topology.SRC: bsm_name,
-                          Topology.DST: node,
-                          Topology.DELAY: args.cc_delay * 1e9})
-
-        cchannels.append({Topology.SRC: node,
-                          Topology.DST: bsm_name,
-                          Topology.DELAY: args.cc_delay * 1e9})
-
+cchannels, qchannels, bsm_nodes = generate_bsm_links(graph, node_procs, args, bsm_name_func)
 nodes += bsm_nodes
 output_dict[Topology.ALL_NODE] = nodes
 output_dict[Topology.ALL_Q_CHANNEL] = qchannels
 
-for node1 in router_names:
-    for node2 in router_names:
-        if node1 == node2:
-            continue
-        cchannels.append({Topology.SRC: node1,
-                          Topology.DST: node2,
-                          Topology.DELAY: args.cc_delay * 1e9})
-
+router_cchannels = generate_classical(router_names, args.cc_delay)
+cchannels += router_cchannels
 output_dict[Topology.ALL_C_CHANNEL] = cchannels
 
-output_dict[Topology.STOP_TIME] = args.stop * 1e12
-if args.parallel:
-    output_dict[RouterNetTopo.IS_PARALLEL] = True
-    output_dict[RouterNetTopo.PROC_NUM] = int(args.parallel[2])
-    output_dict[RouterNetTopo.IP] = args.parallel[0]
-    output_dict[RouterNetTopo.PORT] = int(args.parallel[1])
-    output_dict[RouterNetTopo.LOOKAHEAD] = int(args.parallel[4])
-    if args.parallel[3] == "true":
-        # set all to synchronous
-        output_dict[RouterNetTopo.ALL_GROUP] = \
-            [{RouterNetTopo.TYPE: RouterNetTopo.SYNC} for _ in
-             range(int(args.parallel[2]))]
-    else:
-        output_dict[RouterNetTopo.ALL_GROUP] = \
-            [{RouterNetTopo.TYPE: RouterNetTopo.ASYNC}] * int(args.parallel[2])
-else:
-    output_dict[RouterNetTopo.IS_PARALLEL] = False
+final_config(output_dict, args)
 
 # write final json
 output_file = open(args.output, 'w')
